+++ RandomSkunk.Results.UnitTests/SelectMany_Async_methods.cs
@@ -140,6 +140,17 @@ public class SelectMany_Async_methods
         }
 
         [Fact]
+        public async Task Given_target_is_Maybe_of_T_and_onNoneSelector_is_provided_When_source_is_None_Returns_result_from_onNoneSelector()
+        {
+            var source = Maybe<int>.None;
+
+            var actual = await source.SelectMany(value => Task.FromResult(value.ToString().ToMaybe()), () => Task.FromResult(Maybe<string>.Fail("error from onNoneSelector")));
+
+            actual.IsFail.Should().BeTrue();
+            actual.Error.Message.Should().Be("error from onNoneSelector");
+        }
+
+        [Fact]
         public async Task Given_target_is_Maybe_of_T_and_onSuccess_function_is_null_Throws_ArgumentNullException()
         {
             var source = Maybe<int>.Fail();
@@ -174,6 +185,17 @@ public class SelectMany_Async_methods
             actual.Error.Message.Should().Be(expectedError.Message);
             actual.Error.ErrorCode.Should().Be(expectedError.ErrorCode);
             actual.Error.Title.Should().Be(expectedError.Title);
+        }
+
+        [Fact]
+        public async Task Given_target_is_Result_and_onNoneSelector_is_provided_When_source_is_None_Returns_result_from_onNoneSelector()
+        {
+            var source = Maybe<string>.None;
+
+            var actual = await source.SelectMany(value => Task.FromResult(Result.Success()), () => Task.FromResult(Result.Fail("error from onNoneSelector")));
+
+            actual.IsFail.Should().BeTrue();
+            actual.Error.Message.Should().Be("error from onNoneSelector");
         }
 
         [Fact]
@@ -223,6 +245,17 @@ public class SelectMany_Async_methods
             actual.Error.Message.Should().Be(expectedError.Message);
             actual.Error.ErrorCode.Should().Be(expectedError.ErrorCode);
             actual.Error.Title.Should().Be(expectedError.Title);
+        }
+
+        [Fact]
+        public async Task Given_target_is_Result_of_T_and_onNoneSelector_is_provided_When_source_is_None_Returns_result_from_onNoneSelector()
+        {
+            var source = Maybe<int>.None;
+
+            var actual = await source.SelectMany(value => Task.FromResult(value.ToString().ToResult()), () => Task.FromResult(Result<string>.Fail("error from onNoneSelector")));
+
+            actual.IsFail.Should().BeTrue();
+            actual.Error.Message.Should().Be("error from onNoneSelector");
         }
 
         [Fact]