+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning].
 - RandomSkunk.Results:
     - Change `Maybe.None()` method to property.
 
+### Added
+
+- RandomSkunk.Results:
+    - Add optional `onNoneSelector` parameter to `Maybe<T>.Select` and `Maybe<T>.SelectMany` methods.
+
 ## [1.0.0-alpha23] - 2023-01-03
 
 ### Fixed